# docs/swagger/swagger.yaml
definitions:
  admin.parameters:
    properties:
      action:
        type: string
      file_path:
        type: string
      wechatid:
        items:
          type: string
        type: array
    type: object
  app.ResponseMsg:
    properties:
      data:
        type: object
      msg:
        type: string
      success:
        type: boolean
    type: object
  models.Applicant:
    properties:
      address:
        type: string
      cert_id:
        type: string
      cert_name:
        type: string
      cert_type:
        type: integer
      city:
        type: string
      company_name:
        type: string
      company_nature:
        type: string
      company_scale:
        type: string
      district:
        type: string
      english_name:
        type: string
      job:
        type: string
      name:
        type: string
      personal_id:
        type: string
      phone:
        type: integer
      province:
        type: string
      study_date:
        type: string
      wechat_id:
        type: string
    type: object
  models.C_certs:
    properties:
      authority:
        type: string
      cert_id:
        type: string
      cert_name:
        type: string
      image_design:
        $ref: '#/definitions/models.ImageDesigner'
        type: object
      price:
        type: integer
      status:
        type: string
    type: object
  models.Coord:
    properties:
      font:
        type: string
      font_size:
        type: number
      str:
        type: string
      x:
        type: integer
      "y":
        type: integer
    type: object
  models.ImageDesigner:
    properties:
      date:
        $ref: '#/definitions/models.Coord'
        type: object
      english_name:
        $ref: '#/definitions/models.Coord'
        type: object
      img_name:
        type: string
      name:
        $ref: '#/definitions/models.Coord'
        type: object
      personal_id:
        $ref: '#/definitions/models.Coord'
        type: object
      serial_number:
        $ref: '#/definitions/models.Coord'
        type: object
    type: object
info:
  contact: {}
  description: Certificate authentication system.
  license: {}
  title: TAXCAS Example API
  version: "1.0"
paths:
  /api/auth:
    get:
      parameters:
      - description: The username for login
        in: query
        name: username
        required: true
        type: string
      - description: The password for login
        in: query
        name: password
        required: true
        type: string
      responses:
        "200":
          description: data":{"token":string}
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      summary: 用户登陆
      tags:
      - 认证授权
  /api/v1/admin/applicants/certs/{certid}:
    get:
      description: 查询指定证书的申领信息
      parameters:
      - description: Cert ID
        in: path
        name: certid
        required: true
        type: string
      - description: 类型 all | export | verify | passed | Reject
        in: query
        name: type
        required: true
        type: string
      - description: 页数
        in: query
        name: page
        type: integer
      - description: 每页显示的条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: data:[{""}]
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查询证书申领信息
      tags:
      - 后台管理
    put:
      parameters:
      - description: 'file_path: 导入的csv文件路径, action: passed(审核中) | refunded(已拒绝),
          wechatid: 选中的记录'
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin.parameters'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: data:{""}
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 执行审核结果
      tags:
      - 后台管理
  /api/v1/admin/certs:
    get:
      description: 查询所有证书列表
      produces:
      - application/json
      responses:
        "200":
          description: data:[{"cert_id":"0", "cert_name":"证书1", "status":"enable"}]
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取证书列表
      tags:
      - 后台管理
    post:
      parameters:
      - description: 证书详细信息
        in: body
        name: certInfo
        required: true
        schema:
          $ref: '#/definitions/models.C_certs'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: certID 不需要填写, 失败返回 false 及 msg
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 添加证书
      tags:
      - 后台管理
  /api/v1/admin/excels:
    post:
      parameters:
      - description: 审核结果.csv
        in: formData
        name: excel
        required: true
        type: file
      responses:
        "200":
          description: data:{""}
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 导入文件
      tags:
      - 后台管理
  /api/v1/admin/files/applicants/certs/{certid}:
    get:
      parameters:
      - description: Cert ID
        in: path
        name: certid
        required: true
        type: string
      - description: 类型 export | Reject
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'data:{"file_save_path":"upload/images/96a.csv", "file_url":
            "http://..."}'
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 导出用户申领信息
      tags:
      - 后台管理
  /api/v1/admin/fonts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 'data:{"image_save_path":"upload/images/96a.jpg", "image_url":
            "http://..."}'
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查询字体列表
      tags:
      - 后台管理
  /api/v1/admin/images:
    post:
      parameters:
      - description: 证书模板图片
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 'data:{"image_save_path":"upload/images/96a.jpg", "image_url":
            "http://..."}'
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 上传证书模板
      tags:
      - 后台管理
  /api/v1/admin/images/cert:
    get:
      parameters:
      - description: 证书详细信息
        in: body
        name: positions
        schema:
          $ref: '#/definitions/models.ImageDesigner'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'data:{"image_save_path":"upload/images/96a.jpg", "image_url":
            "http://..."}'
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 预览证书
      tags:
      - 后台管理
  /api/v1/admin/images/certs/{certid}/{wechatid}:
    get:
      parameters:
      - description: 证书id
        in: path
        name: certid
        required: true
        type: string
      - description: 用户微信id
        in: path
        name: wechatid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'data:{"image_save_path":"export/images/96a.jpg", "image_url":
            "http://..."}'
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查看用户证书
      tags:
      - 后台管理
  /api/v1/weixin/applicants/users:
    post:
      parameters:
      - description: 用户提交信息
        in: body
        name: applicant
        required: true
        schema:
          $ref: '#/definitions/models.Applicant'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: cost 与 applyStatus 不提交. 失败返回 false 及 msg
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 申请证书
      tags:
      - 微信公众号
  /api/v1/weixin/certs:
    get:
      description: 查询所有证书列表
      produces:
      - application/json
      responses:
        "200":
          description: data:[{"cert_id":"0", "cert_name":"证书1", "status":"enable"}]
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取证书列表
      tags:
      - 微信公众号
  /api/v1/weixin/images/certs/{certid}/{wechatid}:
    get:
      parameters:
      - description: 证书id
        in: path
        name: certid
        required: true
        type: string
      - description: 用户微信id
        in: path
        name: wechatid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'data:{"image_save_path":"export/images/96a.jpg", "image_url":
            "http://..."}'
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查看用户证书
      tags:
      - 微信公众号
  /api/v1/weixin/openid/{code}:
    get:
      parameters:
      - description: 微信浏览器获取的 code
        in: path
        name: code
        required: true
        type: string
      responses:
        "200":
          description: '{"openid":string, "token":string}'
          schema:
            type: string
      summary: 获取用户openid
      tags:
      - 微信公众号
  /api/v1/weixin/users/{openid}:
    get:
      parameters:
      - description: 用户openid
        in: path
        name: openid
        required: true
        type: string
      responses:
        "200":
          description: '用户基本信息 及 证书申领状态 ["申请证书id" : "申请状态"]'
          schema:
            $ref: '#/definitions/app.ResponseMsg'
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查询用户
      tags:
      - 微信公众号
  /api/v1/weixin/wxorder/{openid}/{certid}:
    get:
      parameters:
      - description: 用户 openid
        in: path
        name: openid
        required: true
        type: string
      - description: 证书 id
        in: path
        name: certid
        required: true
        type: string
      responses:
        "200":
          description: '{"prepay_id":string}'
          schema:
            type: string
      summary: 获取支付订单
      tags:
      - 微信公众号
  /api/v1/weixin/wxquery/{out_trade_no}:
    get:
      parameters:
      - description: 付款订单号
        in: path
        name: out_trade_no
        required: true
        type: string
      responses:
        "200":
          description: '{"msg":string, "extra":}'
          schema:
            type: string
      summary: 查询退款
      tags:
      - 微信公众号
  /api/v1/weixin/wxrefund/{out_trade_no}:
    get:
      parameters:
      - description: 付款订单号
        in: path
        name: out_trade_no
        required: true
        type: string
      responses:
        "200":
          description: '{"msg":string, "extra":}'
          schema:
            type: string
      summary: 申请退款
      tags:
      - 微信公众号
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"